Refactor Navbar
